# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fleet Tracker Lite API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trucks:
    get:
      operationId: listTrucks
      tags: [trucks]
      summary: List all trucks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of trucks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Truck"
    post:
      operationId: createTruck
      tags: [trucks]
      summary: Add a new truck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
  /trucks/{id}:
    get:
      operationId: getTruck
      tags: [trucks]
      summary: Get truck details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Truck details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
        "404":
          description: Not found
    put:
      operationId: updateTruck
      tags: [trucks]
      summary: Update a truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckUpdate"
      responses:
        "200":
          description: Updated truck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
        "404":
          description: Not found
    delete:
      operationId: deleteTruck
      tags: [trucks]
      summary: Delete a truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Truck:
      type: object
      required: [id, registrationNumber, model, capacity, fuelType, status, createdAt]
      properties:
        id:
          type: integer
        registrationNumber:
          type: string
        model:
          type: string
        capacity:
          type: number
        fuelType:
          type: string
        status:
          type: string
          enum: [available, assigned, running, maintenance]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
    TruckInput:
      type: object
      required: [registrationNumber, model, capacity, fuelType]
      properties:
        registrationNumber:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        capacity:
          type: number
        fuelType:
          type: string
          enum: [diesel, petrol, electric, hybrid]
        notes:
          type: string
    TruckUpdate:
      type: object
      properties:
        registrationNumber:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        capacity:
          type: number
        fuelType:
          type: string
          enum: [diesel, petrol, electric, hybrid]
        status:
          type: string
          enum: [available, assigned, running, maintenance]
        notes:
          type: string
...