# lib/api-zod/src/generated/types/truckInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Tracker Lite API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TruckInputFuelType } from './truckInputFuelType';

export interface TruckInput {
  /** @minLength 1 */
  registrationNumber: string;
  /** @minLength 1 */
  model: string;
  capacity: number;
  fuelType: TruckInputFuelType;
  notes?: string;
}